requirements.txt
argparse~=1.4.0
uvloop~=0.15.2
aiohttp~=3.7.4.post0
python-dotenv~=0.19.0
sqlalchemy~=1.4.22
